Reorganize Submission endpoints
