+++ apps/server/swagger/data-api.yml
@@ -221,77 +221,3 @@
         $ref: '#/components/responses/ServerError'
       503:
         $ref: '#/components/responses/ServiceUnavailableError'
-
-/data/{systemId}:
-  delete:
-    summary: Delete Submitted Data by System ID. Returns an Active Submission with the records to be deleted
-    tags:
-      - Admin
-    parameters:
-      - name: systemId
-        in: path
-        required: true
-        schema:
-          type: string
-    responses:
-      200:
-        description: Returns an Active Submission with the records to be deleted
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/DeleteDataActiveSubmission'
-      400:
-        $ref: '#/components/responses/BadRequest'
-      401:
-        $ref: '#/components/responses/UnauthorizedError'
-      404:
-        $ref: '#/components/responses/NotFound'
-      409:
-        $ref: '#/components/responses/StatusConflict'
-      500:
-        $ref: '#/components/responses/ServerError'
-      503:
-        $ref: '#/components/responses/ServiceUnavailableError'
-
-/data/category/{categoryId}/edit:
-  post:
-    summary: Edit Submitted Data. Returns an Active Submission with the records to be updated
-    tags:
-      - Admin
-    consumes:
-      - multipart/form-data
-    parameters:
-      - name: categoryId
-        in: path
-        required: true
-        schema:
-          type: string
-        description: ID of the category
-    requestBody:
-      content:
-        multipart/form-data:
-          schema:
-            type: object
-            properties:
-              files:
-                type: array
-                items:
-                  type: string
-                  format: binary
-              organization:
-                type: string
-    responses:
-      200:
-        description: Edit Data request accepted
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/CreateSubmissionResult'
-      400:
-        $ref: '#/components/responses/BadRequest'
-      401:
-        $ref: '#/components/responses/UnauthorizedError'
-      500:
-        $ref: '#/components/responses/ServerError'
-      503:
-        $ref: '#/components/responses/ServiceUnavailableError'
