+++ apps/server/swagger/tags.yml
@@ -1,6 +1,4 @@
 tags:
-  - name: Admin
-    description: Admin operations to manipulate Submitted Data
   - name: Audit
     description: Retrieves a list of changes made to Submitted Data records
   - name: Category
