+++ packages/data-provider/src/controllers/submittedDataController.ts
@@ -3,13 +3,10 @@ import * as _ from 'lodash-es';
 import { BaseDependencies } from '../config/config.js';
 import submittedDataService from '../services/submittedDataService.js';
 import { parseSQON } from '../utils/convertSqonToQuery.js';
-import { BadRequest, NotFound } from '../utils/errors.js';
-import { processFiles } from '../utils/fileUtils.js';
+import { NotFound } from '../utils/errors.js';
 import { asArray } from '../utils/formatUtils.js';
 import { validateRequest } from '../utils/requestValidation.js';
 import {
-	dataDeleteBySystemIdRequestSchema,
-	dataEditRequestSchema,
 	dataGetByCategoryRequestSchema,
 	dataGetByOrganizationRequestSchema,
 	dataGetByQueryRequestschema,
@@ -23,73 +20,6 @@ const controller = (dependencies: BaseDependencies) => {
 	const defaultPage = 1;
 	const defaultPageSize = 20;
 	return {
-		deleteSubmittedDataBySystemId: validateRequest(dataDeleteBySystemIdRequestSchema, async (req, res, next) => {
-			try {
-				const systemId = req.params.systemId;
-
-				logger.info(LOG_MODULE, `Request Delete Submitted Data systemId '${systemId}'`);
-
-				// TODO: get userName from auth
-				const userName = '';
-
-				const deletedRecords = await service.deleteSubmittedDataBySystemId(systemId, userName);
-
-				const response = {
-					submissionId: deletedRecords.submissionId,
-					records: deletedRecords.data,
-				};
-
-				return res.status(200).send(response);
-			} catch (error) {
-				next(error);
-			}
-		}),
-
-		editSubmittedData: validateRequest(dataEditRequestSchema, async (req, res, next) => {
-			try {
-				const categoryId = Number(req.params.categoryId);
-				const files = req.files as Express.Multer.File[];
-				const organization = req.body.organization;
-
-				logger.info(LOG_MODULE, `Request Edit Submitted Data`);
-
-				// TODO: get userName from auth
-				const userName = '';
-
-				if (!files || files.length == 0) {
-					throw new BadRequest(
-						'The "files" parameter is missing or empty. Please include files in the request for processing.',
-					);
-				}
-
-				const { validFiles, fileErrors } = await processFiles(files);
-
-				if (fileErrors.length == 0) {
-					logger.info(LOG_MODULE, `File uploaded successfully`);
-				} else {
-					logger.info(LOG_MODULE, 'Found some errors processing this request');
-				}
-
-				const editSubmittedDataResult = await service.editSubmittedData({
-					files: validFiles,
-					categoryId,
-					organization,
-					userName,
-				});
-
-				const response = {
-					status: editSubmittedDataResult.status,
-					submissionId: editSubmittedDataResult.submissionId,
-					inProcessEntities: editSubmittedDataResult.inProcessEntities,
-					batchErrors: [...fileErrors, ...editSubmittedDataResult.batchErrors],
-				};
-
-				return res.status(200).send(response);
-			} catch (error) {
-				next(error);
-			}
-		}),
-
 		getSubmittedDataByCategory: validateRequest(dataGetByCategoryRequestSchema, async (req, res, next) => {
 			try {
 				const categoryId = Number(req.params.categoryId);
