+++ packages/data-provider/src/routers/submittedDataRouter.ts
@@ -1,20 +1,12 @@
-import { json, Router, urlencoded } from 'express';
-import multer from 'multer';
+import { Router, urlencoded } from 'express';
 
 import { BaseDependencies } from '../config/config.js';
 import submittedDataController from '../controllers/submittedDataController.js';
 import { auth } from '../middleware/auth.js';
-import { getSizeInBytes } from '../utils/fileUtils.js';
 
 const router = (dependencies: BaseDependencies): Router => {
 	const router = Router();
 	router.use(urlencoded({ extended: false }));
-
-	const fileSizeLimit = getSizeInBytes(dependencies.limits.fileSize);
-
-	const upload = multer({ dest: '/tmp', limits: { fileSize: fileSizeLimit } });
-
-	router.use(json({ limit: fileSizeLimit }));
 
 	router.get('/category/:categoryId', auth, submittedDataController(dependencies).getSubmittedDataByCategory);
 
@@ -28,15 +20,6 @@ const router = (dependencies: BaseDependencies): Router => {
 		'/category/:categoryId/organization/:organization/query',
 		auth,
 		submittedDataController(dependencies).getSubmittedDataByQuery,
-	);
-
-	router.delete(`/:systemId`, auth, submittedDataController(dependencies).deleteSubmittedDataBySystemId);
-
-	router.post(
-		`/category/:categoryId/edit`,
-		auth,
-		upload.array('files'),
-		submittedDataController(dependencies).editSubmittedData,
 	);
 
 	return router;
